Restructure the source code
