+++ artifacts/nft/open/NFTOpenCollection.ral.json
