+++ artifacts/nft/open/NFTOpenCollectionWithRoyalty.ral.json
