+++ artifacts/nft/publicsale/random/NFTPublicSaleCollectionRandom.ral.json
