+++ artifacts/nft/publicsale/random/NFTPublicSaleCollectionRandomWithRoyalty.ral.json
