+++ artifacts/nft/publicsale/sequential/NFTPublicSaleCollectionSequentialWithRoyalty.ral.json
