+++ contracts/nft/open/nft_open_collection_with_royalty.ral
